refactor(issue-23): reorganize check test suites
