+++ tests/checks/suites/__init__.py
@@ -1,0 +1,1 @@
+# Package marker for tests.checks.suites helper module imports.
