+++ tests/checks/suites/core/test_clang_tidy_check.py
