+++ tests/checks/suites/policy/test_fmea_action_register.py
