+++ tests/checks/suites/policy/test_test_catalog.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 from python_tools.core.models import CheckContext
 from python_tools.policies.test_catalog import TestCatalogCheck
-from tests.checks.tests.path_specs import TESTS_UNIT_DIR, cpp_file
+from tests.checks.suites.support.path_specs import TESTS_UNIT_DIR, cpp_file
 
 
 def _write(path: Path, content: str) -> None:
