+++ tests/checks/suites/release/test_tool_manifest.py
