--- tests/checks/tests/__init__.py
@@ -1,1 +1,0 @@
-# Package marker for tests.checks.tests helper module imports.
